IVABOT/requirements.txt
python-telegram-bot[job-queue]>=20.0
httpx
beautifulsoup4
aiosqlite
cryptography
